fix(rust-codegen): drive ternary/literal widths from the destination; never emit i32-defaulted ifs
Two blocking findings from the PR review (each validated end-to-end
with the branch's own compactc 0.31.119 + rustc 1.96 against the base
branch, which refused these programs cleanly):

1. Ledger cell alignment divergence (silent state-byte parity bug).
   coerce-literal-if-rhs-rendered sized literal ternary arms from the
   const's own inferred type, so `const picked = hot ? 10 : 20` written
   to `lastPick: Uint<64>` committed a 1-byte-aligned cell where the TS
   backend's field descriptor (CompactTypeUnsignedInteger(2^64-1, 8))
   commits 8 — same decoded value, divergent serialized state. Both
   cell-write emitters (cell-write-op-lines and emit-body-mutations'
   inline clause) now render plain-Uint writes as
   `new_cell((<val>) as <field-width>)` from current-ledger-field-types,
   mirroring the TS field-descriptor conversion; subtyping makes the
   cast a lossless zero-extension or a no-op.

2. Pure-circuit const ternaries emitted Rust that dies in rustc
   (regression from the base branch's clean refusals). Three fixes:
   - stmt-pure-body-rust's const clause now applies the same
     decl-type literal coercions the impure/constructor const route
     uses (`const p: Field = c ? 1 : 0` -> Fr::from arms; statement
     consts get width suffixes).
   - seq-stmt-rust's `(=)` clause (the let*-lifted route, whose IR
     node carries no binder type) coerces both-literal if arms at
     max(arm) width so a `5000000000 : 0` join cannot default to i32
     and overflow.
   - expr-rust's seq clause inlines the degenerate `(seq (= v
     <lit-if>) v)` at its tail — byte-identical to the no-const
     return position (`pick`), letting the use context size the arms
     instead of stranding the binding at its minimal width (E0308) or
     i32 (overflow).

Regression coverage:
- record_literal_pick_byte_parity: executes recordLiteralPick(true)
  on BOTH backends from the same post-init state and byte-compares
  the serialized envelopes (capture script + TS reference extended;
  verified it fails with the 1-vs-8-byte alignment divergence when
  the cast is removed).
- ternary fixture grows literalPick / bigLiteralPick (the reviewer's
  exact repro shapes) with an executing gate pinning both arm values.
- 9 fixtures regenerate (drift = exactly the new casts/suffixes);
  codegen_regression green, full tests-e2e-rust 126/126, cargo fmt +
  clippy -D warnings clean, reviewer repro shapes type-checked with
  rustc 1.96.

+++ tests-e2e-rust/contracts/bounded-uint-fixture/lib.rs
@@ -79,7 +79,7 @@ where
     ) -> Result<CircuitResults<PS, ()>, CompactError> {
         let ops = OpProgramVerify::<DefaultDB>::new()
             .push(false, new_cell(0u8))
-            .push(true, new_cell(v))
+            .push(true, new_cell((v) as u8))
             .ins(false, 1)
             .build();
